Replication: migrate entities to refactored framework API
EntityFactory -> EntityRegistry; unified OnSerializeConstruction(bs,write) / SerializeFields(FieldSerializer&) / SerializeForcedState(bs,write); ownerGUID is now PeerGuid (convert at call sites).

+++ code/client/src/core/application.cpp
@@ -405,7 +405,7 @@ namespace MafiaMP::Core {
     }
 
     uint64_t Application::GetLocalPlayerOwnerID() {
-        return _localPlayer ? _localPlayer->ownerGUID : 0;
+        return _localPlayer ? static_cast<uint64_t>(_localPlayer->ownerGUID) : 0;
     }
 
     void Application::InitRPCs() {
